gemini_image_generator.py
"""
Gemini API integration for image generation
"""

import google.generativeai as genai
import os
from typing import Optional
import base64
import io
from PIL import Image


class GeminiImageGenerator:
    """Generate images using Google Gemini API"""
    
    def __init__(self, api_key: str, model_name: str = "gemini-1.5-pro"):
        """
        Initialize Gemini image generator
        
        Args:
            api_key: Google API key (OAuth client ID or API key)
            model_name: Model to use (gemini-1.5-pro, gemini-1.5-flash, etc.)
        """
        self.api_key = api_key
        self.model_name = model_name
        genai.configure(api_key=api_key)
        
        # Initialize the model
        try:
            # For image generation, we might need to use text-to-image model
            # Note: Gemini models are primarily text models
            # For actual image generation, you might need Imagen API or DALL-E
            # This is a placeholder that uses Gemini's capabilities
            self.model = genai.GenerativeModel(model_name)
        except Exception as e:
            print(f"Warning: Could not initialize Gemini model: {e}")
            self.model = None
    
    def generate_image_from_text(self, prompt: str, headline: str = "") -> Optional[Image.Image]:
        """
        Generate an image based on text description
        
        IMPORTANT: Gemini models are TEXT models and DO NOT generate images.
        They can only generate text descriptions.
        
        For actual image generation, you need:
        - Google Imagen API (separate service)
        - OpenAI DALL-E API
        - Stable Diffusion API
        - Other image generation services
        
        This function is a placeholder that uses Gemini to create image descriptions,
        which could then be used with an actual image generation API.
        """
        if not self.model:
            return None
        
        try:
            # Create a detailed prompt for image description (not generation)
            full_prompt = f"""
            Create a detailed, professional image description for a news article.
            Headline: {headline}
            Description: {prompt}
            
            The image description should be:
            - Professional and newsworthy
            - Relevant to the headline
            - Suitable for a news website
            - Detailed enough for image generation
            """
            
            # Generate description (Gemini can do this)
            response = self.model.generate_content(full_prompt)
            
            # Note: This returns a TEXT description, not an image
            # To actually generate an image, you would need to:
            # 1. Take this description
            # 2. Send it to an image generation API (Imagen, DALL-E, etc.)
            # 3. Get the generated image back
            print(f"Generated image description: {response.text}")
            
            return None  # Returns None - Gemini doesn't generate images
            
        except Exception as e:
            print(f"Error generating image description: {e}")
            return None
    
    def generate_image_url(self, headline: str, description: str = "") -> Optional[str]:
        """
        Generate image URL or path
        Returns a placeholder URL for now
        """
        # This is a placeholder - in a real implementation, you would:
        # 1. Call an image generation API (Imagen, DALL-E, Stable Diffusion)
        # 2. Save the image
        # 3. Return the URL/path
        
        # For now, return None to use original images
        return None


# Alternative: Use a text-to-image API that actually generates images
def generate_news_image(headline: str, description: str = "", api_key: str = "") -> Optional[str]:
    """
    Generate news image using available APIs
    
    Since Gemini doesn't generate images directly, this is a placeholder
    for integrating with actual image generation services like:
    - Google Imagen API
    - OpenAI DALL-E
    - Stable Diffusion API
    - Midjourney API
    """
    # Placeholder implementation
    # In production, integrate with an actual image generation service
    return None

