requirements-prod.txt
# Production Dependencies for Nepal Times News Platform

# Web Framework
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.12.1

# Scraping & Parsing
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
lxml==4.9.3
html5lib==1.1
urllib3==2.1.0
urllib.robotparser  # Built-in

# Language Detection & NLP
langdetect==1.0.9
polyglot==16.7.4  # Optional, for better language detection
nltk==3.8.1

# Image Processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
scikit-image==0.22.0  # For advanced image processing

# AI/LLM (choose based on provider)
openai==1.3.7  # For GPT models
anthropic==0.7.8  # For Claude
google-generativeai==0.3.2  # For Gemini

# Scheduling
APScheduler==3.10.4
celery==5.3.4  # Optional, for distributed task queue
redis==5.0.1  # For Celery broker and caching

# Caching
redis==5.0.1
python-redis-cache==3.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
chardet==5.2.0  # Encoding detection
tqdm==4.66.1  # Progress bars

# Logging & Monitoring
python-json-logger==2.0.7
structlog==23.2.0
sentry-sdk==1.38.0  # Error tracking

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing (development)
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# News fetching (existing)
newspaper3k==0.2.8
news-please==1.6.13

